requirements.txt
crewai>=0.11.0
pydantic>=2.5.2
pyyaml>=6.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.2
requests>=2.31.0 